feat: creat, get user, info, change passw

+++ src/core/service/auth_service.go
@@ -10,7 +10,9 @@ import (
 	"TTCS/src/core/dto"
 	"TTCS/src/present/httpui/request"
 	"context"
+	"errors"
 	"fmt"
+	"time"
 )
 
 type AuthService struct {
@@ -18,23 +20,43 @@ type AuthService struct {
 	otpRepo      domain.OtpRepo
 	hashProvider crypto.HashProvider
 	otpProvider  crypto.OTPProvider
+	jwtProvider  crypto.JwtProvider
 	mailService  *mail.GmailService
 }
 
-func NewAuthService(userRepo domain.UserRepo, otpRepo domain.OtpRepo, hashProvider crypto.HashProvider, otpProvider crypto.OTPProvider, mailService *mail.GmailService) *AuthService {
+func NewAuthService(userRepo domain.UserRepo, otpRepo domain.OtpRepo, hashProvider crypto.HashProvider, otpProvider crypto.OTPProvider, jwtProvider crypto.JwtProvider, mailService *mail.GmailService) *AuthService {
 	return &AuthService{
 		userRepo:     userRepo,
 		otpRepo:      otpRepo,
 		hashProvider: hashProvider,
 		otpProvider:  otpProvider,
+		jwtProvider:  jwtProvider,
 		mailService:  mailService,
 	}
 }
 
 func (s *AuthService) GenOTP(ctx context.Context, email string) (string, error) {
-	_ = "AuthService.GenOTP"
-	_ = s.otpRepo.DeleteByEmail(ctx, email)
-	otp := s.otpProvider.GenerateCode()
+	caller := "AuthService.GenOTP"
+	user, err := s.userRepo.GetByEmail(ctx, email)
+	if err != nil {
+		return "", err
+	}
+	if user != nil {
+		err := errors.New("user already exists")
+		return "", fault.Wrapf(err, "[%v] failed to create OTP", caller)
+	}
+	otpDb, err := s.otpRepo.GetByEmail(ctx, email)
+	if err != nil {
+		return "", err
+	}
+	if otpDb != nil {
+		if otpDb.CreatedAt.After(time.Now().Add(-5 * time.Minute)) {
+			err := errors.New("otp already exists")
+			return "", fault.Wrapf(err, "[%v] failed to create OTP", caller)
+		}
+		_ = s.otpRepo.DeleteByEmail(ctx, email)
+	}
+	otp := s.otpProvider.GenerateOTP()
 	if err := s.otpRepo.Create(ctx, &domain.Otp{
 		Email: email,
 		Otp:   otp,
@@ -109,11 +131,11 @@ func (s *AuthService) generateToken(ctx context.Context, user *domain.User) (*dt
 		Username: user.Email,
 	}
 
-	accessToken, err := crypto.Generate(configs.GetConfig().Jwt.AccessSecret, payload, configs.GetConfig().Jwt.ExpireAccess)
+	accessToken, err := s.jwtProvider.Generate(configs.GetConfig().Jwt.AccessSecret, payload, configs.GetConfig().Jwt.ExpireAccess)
 	if err != nil {
 		return nil, fault.Wrapf(err, "[%v] failed to gen token", caller).SetTag(fault.TagInternalServer)
 	}
-	refreshToken, err := crypto.Generate(configs.GetConfig().Jwt.RefreshSecret, payload, configs.GetConfig().Jwt.ExpireRefresh)
+	refreshToken, err := s.jwtProvider.Generate(configs.GetConfig().Jwt.RefreshSecret, payload, configs.GetConfig().Jwt.ExpireRefresh)
 	if err != nil {
 		return nil, fault.Wrapf(err, "[%v] failed to gen token", caller).SetTag(fault.TagInternalServer)
 	}
@@ -136,4 +158,46 @@ func (s *AuthService) verifyOTP(ctx context.Context, otp string, email string) (
 		return false, fault.Wrapf(err, "[%v] wrong Otp ", caller).SetTag(fault.TagUnAuthorize)
 	}
 	return true, nil
+}
+
+func (s *AuthService) VerifyToken(ctx context.Context, token string) (*domain.User, error) {
+	caller := "AuthUserUseCase.VerifyToken"
+
+	payload, err := s.jwtProvider.Verify(configs.GetConfig().Jwt.AccessSecret, token)
+	if errors.Is(err, crypto.ErrTokenExpired) {
+		return nil, fault.Wrapf(err, "[%v] token expired", caller).SetTag(fault.TagUnAuthorize)
+
+	}
+	if err != nil {
+		return nil, fault.Wrapf(err, "[%v] failed to verify token", caller).SetTag(fault.TagUnAuthorize)
+	}
+	username := payload.Username
+	user, err := s.userRepo.GetByEmail(ctx, username)
+	if err != nil {
+		return nil, err
+	}
+	return user, nil
+}
+
+func (s *AuthService) ChangePassword(ctx context.Context, req request.ChangePasswordRequest) error {
+	caller := "AuthService.ChangePassword"
+
+	user, err := s.userRepo.GetByEmail(ctx, req.Email)
+	if err != nil {
+		return err
+	}
+	err = s.hashProvider.ComparePassword(req.OldPassword, user.PasswordHash)
+	if err != nil {
+		return fault.Wrapf(err, "[%v] old password is wrong", caller).SetTag(fault.TagUnAuthorize)
+	}
+	hashedPw, err := s.hashProvider.Hash(req.NewPassword)
+	if err != nil {
+		return fault.Wrapf(err, "[%v] failed to hash password", caller).SetTag(fault.TagInternalServer)
+	}
+	user.PasswordHash = hashedPw
+	user, err = s.userRepo.Update(ctx, user)
+	if err != nil {
+		return err
+	}
+	return nil
 }
